+++ dessia_common/workflow/core.py
@@ -389,17 +389,18 @@ class Workflow(Block):
 
         self.block_selectors = {}
         for i, block in enumerate(self.blocks):
-            if block in self.display_blocks + self.export_blocks:
-                name = block.name
-                if not name:
+            name = block.name
+            if not name:
+                if block in self.display_blocks:
                     name = block.type_
-                selector = f"{name} ({i})"
-                self.block_selectors[block] = selector
-            else:
-                self.block_selectors[block] = None
+                elif block in self.export_blocks:
+                    name = block.extension
+                else:
+                    name = "Block"
+            self.block_selectors[block] = f"{name} ({i})"
 
-        self.branch_by_display_selector = self.display_branches
-        self.branch_by_export_format = self.export_branches
+        self.branch_by_display_selector = self.branch_by_selector(self.display_blocks)
+        self.branch_by_export_format = self.branch_by_selector(self.export_blocks)
 
     def handle_pipe(self, pipe):
         """
@@ -542,6 +543,17 @@ class Workflow(Block):
 
         return copied_pipes
 
+    def branch_by_selector(self, blocks: List[Block]):
+        """
+        Return the corresponding branch to each display or export selector
+        """
+        selector_branches = {}
+        for block in blocks:
+            branch = self.secondary_branch_blocks(block)
+            selector = self.block_selectors[block]
+            selector_branches[selector] = branch
+        return selector_branches
+
     @property
     def display_blocks(self):
         """
@@ -563,7 +575,8 @@ class Workflow(Block):
                 display_settings.append(settings)
         return display_settings
 
-    def display_settings(self) -> List[DisplaySetting]:
+    @staticmethod
+    def display_settings() -> List[DisplaySetting]:
         """
         Computes the displays settings of the workflow
         """
@@ -591,30 +604,6 @@ class Workflow(Block):
                 format_["selector"] = self.block_selectors[block]
                 export_formats.append(format_)
         return export_formats
-
-    @property
-    def display_branches(self):
-        """
-        Return the corresponding branch to each display selector
-        """
-        selector_branches = {}
-        for display_block in self.display_blocks:
-            branch = self.secondary_branch_blocks(display_block)
-            selector = self.block_selectors[display_block]
-            selector_branches[selector] = branch
-        return selector_branches
-
-    @property
-    def export_branches(self):
-        """
-        Return the corresponding branch to each export format
-        """
-        format_branches = {}
-        for export_block in self.export_blocks:
-            branch = self.secondary_branch_blocks(export_block)
-            selector = self.block_selectors[export_block]
-            format_branches[selector] = branch
-        return format_branches
 
     def _export_formats(self):
         """
@@ -1402,7 +1391,7 @@ class Workflow(Block):
         imports = []
         imports_as_is = []
         for iblock, block in enumerate(self.blocks):
-            block_script = block._to_script()
+            block_script = block._to_script(prefix)
             imports.extend(block_script.imports)
             if block_script.before_declaration is not None:
                 blocks_str += f"{block_script.before_declaration}\n"
@@ -1418,6 +1407,9 @@ class Workflow(Block):
             nbvs_str += f"{prefix}variable_{nbv_index} = {nbv_script.declaration}\n"
 
         # --- Pipes ---
+        if len(self.pipes) > 0:
+            imports.append(self.pipes[0].full_classname)
+
         pipes_str = ""
         for ipipe, pipe in enumerate(self.pipes):
             input_index = self.variable_indices(pipe.input_variable)
@@ -1463,8 +1455,6 @@ class Workflow(Block):
 
         self_script = self._to_script()
         self_script.imports.append(self.full_classname)
-        if len(self.pipes) > 0:
-            self_script.imports.append(self.pipes[0].full_classname)
 
         script_imports = self_script.imports_to_str()
 
@@ -2037,10 +2027,16 @@ class WorkflowState(DessiaObject):
         Reads block to compute available export formats
         """
         export_formats = DessiaObject._export_formats(self)
-        for i, block in enumerate(self.workflow.blocks):
-            if hasattr(block, "_export_format"):
-                export_formats.append(block._export_format(i))
+
+        # Exportable Blocks
+        export_formats.extend(self.workflow.blocks_export_formats)
         return export_formats
+
+    def export_format_from_selector(self, selector):
+        for export_format in self.workflow.blocks_export_formats:
+            if export_format["selector"] == selector:
+                return export_format
+        raise ValueError(f"No block defines an export with the selector '{selector}'")
 
     def export(self, stream: Union[BinaryFile, StringFile], block_index: int):
         """
@@ -2048,9 +2044,7 @@ class WorkflowState(DessiaObject):
         """
         block = self.workflow.blocks[block_index]
         selector = self.workflow.block_selectors[block]
-        export_format = self.workflow.branch_by_export_format[selector]
-        # export_format = block._export_format(block_index)
-        branch = self.workflow.branch_by_export_format[export_format['export_name']]
+        branch = self.workflow.branch_by_export_format[selector]
         evaluated_blocks = self.evaluate_branch(branch)
         if block not in evaluated_blocks:
             msg = f"Could not reach block at index {block_index}." \