+++ tests/workflow/workflow_sampling.py
@@ -1,12 +1,12 @@
 from dessia_common.tests import RandDataD2
-from dessia_common.datatools.sampling import Sampler
+from dessia_common.datatools.sampling import ClassSampler
 from dessia_common.optimization import FixedAttributeValue, BoundedAttributeValue
 from dessia_common.workflow.blocks import InstantiateModel, ModelMethod
 from dessia_common.typings import MethodType
 from dessia_common.workflow.core import Workflow, Pipe
 
-block_0 = InstantiateModel(model_class=Sampler, name='Sampler')
-block_1 = ModelMethod(method_type=MethodType(Sampler, 'make_doe'), name='gytgy')
+block_0 = InstantiateModel(model_class=ClassSampler, name='ClassSampler')
+block_1 = ModelMethod(method_type=MethodType(ClassSampler, 'make_doe'), name='Make DOE method')
 blocks = [block_0, block_1]
 
 pipe_0 = Pipe(block_0.outputs[0], block_1.inputs[0])