+++ CHANGELOG.md
@@ -7,11 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.10.2
+
+- non serializable attributes were not working
+- wrong import to sklearn -> scikit-learn
+
+## v0.10.1
+
+### Fixed
+- jsonschema bug
+- time rendering on workflow
+
+## v0.10.0
+
 ### Added
 - Generic save to file
 - Sampler class
 - (De)Serialization handles 'typing.Type'
 - Workflow: handle position in to_dict / dict_to_object process
+- sub_matrix method to Dataset
 
 
 ### Breaking Changes
@@ -25,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * class Sampler is now coded in file sampling.py
     * Metrics function are now coded in file metrics.py
 - Retrocompatibility is supported for the present time, with a big warning
+- pareto methods of datatools are now called with a list of attributes and not anymore a costs matrix
 
 
 ### Changes
@@ -34,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - switch is_jsonable to orjson
 
 ### Fixed
+- Workflow: improve to_script
 
 
 ## v0.10.0 [9/26/2022]